Revert "feat: send files and images in chat box"
