+++ app/Http/Controllers/Api/ChatController.php
@@ -551,137 +551,6 @@ class ChatController extends Controller
     }
 
     /**
-     * Upload and send a file
-     */
-    public function uploadFile(Request $request, Conversation $conversation)
-    {
-        $request->validate([
-            'file' => 'required|file|max:20480', // 20MB max
-            'caption' => 'sometimes|string|max:500'
-        ]);
-
-        try {
-            $user = Auth::user();
-
-            if (!$conversation->participants->contains($user->id)) {
-                return response()->json([
-                    'success' => false,
-                    'message' => 'Access denied'
-                ], 403);
-            }
-
-            $file = $request->file('file');
-            $filename = time() . '_' . $file->getClientOriginalName();
-            $filePath = $file->storeAs('chat/files', $filename, 'public');
-            
-            // Get file information
-            $fileSize = $file->getSize();
-            $mimeType = $file->getMimeType();
-            $originalName = $file->getClientOriginalName();
-
-            // Create message with file metadata
-            $message = $conversation->messages()->create([
-                'user_id' => $user->id,
-                'content' => $request->caption ?? "Sent a file: {$originalName}",
-                'type' => 'file',
-                'metadata' => [
-                    'file_path' => $filePath,
-                    'file_name' => $originalName,
-                    'file_size' => $fileSize,
-                    'mime_type' => $mimeType,
-                    'file_url' => asset('storage/' . $filePath)
-                ]
-            ]);
-
-            $message->load('user');
-
-            // Update conversation timestamp
-            $conversation->touch();
-
-            // Broadcast the message
-            broadcast(new MessageSent($message, $conversation))->toOthers();
-
-            return response()->json([
-                'success' => true,
-                'data' => ['message' => $message]
-            ], 201);
-
-        } catch (\Exception $e) {
-            Log::error('Error uploading file: ' . $e->getMessage());
-            return response()->json([
-                'success' => false,
-                'message' => 'Failed to upload file'
-            ], 500);
-        }
-    }
-
-    /**
-     * Upload and send an image
-     */
-    public function uploadImage(Request $request, Conversation $conversation)
-    {
-        $request->validate([
-            'image' => 'required|image|mimes:jpeg,png,jpg,gif,webp|max:10240', // 10MB max
-            'caption' => 'sometimes|string|max:500'
-        ]);
-
-        try {
-            $user = Auth::user();
-
-            if (!$conversation->participants->contains($user->id)) {
-                return response()->json([
-                    'success' => false,
-                    'message' => 'Access denied'
-                ], 403);
-            }
-
-            $image = $request->file('image');
-            $filename = time() . '_' . $image->getClientOriginalName();
-            $imagePath = $image->storeAs('chat/images', $filename, 'public');
-            
-            // Get image information
-            $imageSize = $image->getSize();
-            $dimensions = getimagesize($image);
-            $originalName = $image->getClientOriginalName();
-
-            // Create message with image metadata
-            $message = $conversation->messages()->create([
-                'user_id' => $user->id,
-                'content' => $request->caption ?? '',
-                'type' => 'image',
-                'metadata' => [
-                    'image_path' => $imagePath,
-                    'image_name' => $originalName,
-                    'image_size' => $imageSize,
-                    'width' => $dimensions[0] ?? null,
-                    'height' => $dimensions[1] ?? null,
-                    'image_url' => asset('storage/' . $imagePath)
-                ]
-            ]);
-
-            $message->load('user');
-
-            // Update conversation timestamp
-            $conversation->touch();
-
-            // Broadcast the message
-            broadcast(new MessageSent($message, $conversation))->toOthers();
-
-            return response()->json([
-                'success' => true,
-                'data' => ['message' => $message]
-            ], 201);
-
-        } catch (\Exception $e) {
-            Log::error('Error uploading image: ' . $e->getMessage());
-            return response()->json([
-                'success' => false,
-                'message' => 'Failed to upload image'
-            ], 500);
-        }
-    }
-
-    /**
      * Helper method to mark conversation as read
      */
     private function markConversationAsRead(Conversation $conversation, User $user)
